# [FLINK-10742][network] Let Netty use Flink's buffers directly on receiver side