--- a/flink-runtime/src/main/java/org/apache/flink/runtime/io/network/netty/NettyMessage.java
+++ b/flink-runtime/src/main/java/org/apache/flink/runtime/io/network/netty/NettyMessage.java
@@ -228,17 +228,11 @@ public abstract class NettyMessage {
 
 				final NettyMessage decodedMsg;
 				switch (msgId) {
-					case BufferResponse.ID:
-						decodedMsg = BufferResponse.readFrom(msg);
-						break;
 					case PartitionRequest.ID:
 						decodedMsg = PartitionRequest.readFrom(msg);
 						break;
 					case TaskEventRequest.ID:
 						decodedMsg = TaskEventRequest.readFrom(msg, getClass().getClassLoader());
-						break;
-					case ErrorResponse.ID:
-						decodedMsg = ErrorResponse.readFrom(msg);
 						break;
 					case CancelPartitionRequest.ID:
 						decodedMsg = CancelPartitionRequest.readFrom(msg);
@@ -269,9 +263,12 @@ public abstract class NettyMessage {
 
 	static class BufferResponse extends NettyMessage {
 
-		private static final byte ID = 0;
+		static final byte ID = 0;
 
-		final ByteBuf buffer;
+		// receiver ID (16), sequence number (4), backlog (4), isBuffer (1), isCompressed (1), buffer size (4)
+		static final int MESSAGE_HEADER_LENGTH = 16 + 4 + 4 + 1 + 1 + 4;
+
+		final Buffer buffer;
 
 		final InputChannelID receiverId;
 
@@ -283,19 +280,23 @@ public abstract class NettyMessage {
 
 		final boolean isCompressed;
 
+		final int bufferSize;
+
 		private BufferResponse(
-				ByteBuf buffer,
+				@Nullable Buffer buffer,
 				boolean isBuffer,
 				boolean isCompressed,
 				int sequenceNumber,
 				InputChannelID receiverId,
-				int backlog) {
-			this.buffer = checkNotNull(buffer);
+				int backlog,
+				int bufferSize) {
+			this.buffer = buffer;
 			this.isBuffer = isBuffer;
 			this.isCompressed = isCompressed;
 			this.sequenceNumber = sequenceNumber;
 			this.receiverId = checkNotNull(receiverId);
 			this.backlog = backlog;
+			this.bufferSize = bufferSize;
 		}
 
 		BufferResponse(
@@ -303,24 +304,28 @@ public abstract class NettyMessage {
 				int sequenceNumber,
 				InputChannelID receiverId,
 				int backlog) {
-			this.buffer = checkNotNull(buffer).asByteBuf();
+			this.buffer = checkNotNull(buffer);
 			this.isBuffer = buffer.isBuffer();
 			this.isCompressed = buffer.isCompressed();
 			this.sequenceNumber = sequenceNumber;
 			this.receiverId = checkNotNull(receiverId);
 			this.backlog = backlog;
+			this.bufferSize = buffer.getSize();
 		}
 
 		boolean isBuffer() {
 			return isBuffer;
 		}
 
-		ByteBuf getNettyBuffer() {
+		@Nullable
+		public Buffer getBuffer() {
 			return buffer;
 		}
 
 		void releaseBuffer() {
-			buffer.release();
+			if (buffer != null) {
+				buffer.recycleBuffer();
+			}
 		}
 
 		// --------------------------------------------------------------------
@@ -329,18 +334,13 @@ public abstract class NettyMessage {
 
 		@Override
 		ByteBuf write(ByteBufAllocator allocator) throws IOException {
-			// receiver ID (16), sequence number (4), backlog (4), isBuffer (1), isCompressed (1), buffer size (4)
-			final int messageHeaderLength = 16 + 4 + 4 + 1 + 1 + 4;
-
 			ByteBuf headerBuf = null;
 			try {
-				if (buffer instanceof Buffer) {
-					// in order to forward the buffer to netty, it needs an allocator set
-					((Buffer) buffer).setAllocator(allocator);
-				}
+				// in order to forward the buffer to netty, it needs an allocator set
+				buffer.setAllocator(allocator);
 
 				// only allocate header buffer - we will combine it with the data buffer below
-				headerBuf = allocateBuffer(allocator, ID, messageHeaderLength, buffer.readableBytes(), false);
+				headerBuf = allocateBuffer(allocator, ID, MESSAGE_HEADER_LENGTH, bufferSize, false);
 
 				receiverId.writeTo(headerBuf);
 				headerBuf.writeInt(sequenceNumber);
@@ -351,38 +351,67 @@ public abstract class NettyMessage {
 
 				CompositeByteBuf composityBuf = allocator.compositeDirectBuffer();
 				composityBuf.addComponent(headerBuf);
-				composityBuf.addComponent(buffer);
+				composityBuf.addComponent(buffer.asByteBuf());
 				// update writer index since we have data written to the components:
-				composityBuf.writerIndex(headerBuf.writerIndex() + buffer.writerIndex());
+				composityBuf.writerIndex(headerBuf.writerIndex() + buffer.asByteBuf().writerIndex());
 				return composityBuf;
 			}
 			catch (Throwable t) {
 				if (headerBuf != null) {
 					headerBuf.release();
 				}
-				buffer.release();
+				buffer.recycleBuffer();
 
 				ExceptionUtils.rethrowIOException(t);
 				return null; // silence the compiler
 			}
 		}
 
-		static BufferResponse readFrom(ByteBuf buffer) {
-			InputChannelID receiverId = InputChannelID.fromByteBuf(buffer);
-			int sequenceNumber = buffer.readInt();
-			int backlog = buffer.readInt();
-			boolean isBuffer = buffer.readBoolean();
-			boolean isCompressed = buffer.readBoolean();
-			int size = buffer.readInt();
+		/**
+		 * Parses the message header part and composes a new BufferResponse with an empty data buffer. The
+		 * data buffer will be filled in later.
+		 *
+		 * @param messageHeader the serialized message header.
+		 * @param bufferAllocator the allocator for network buffer.
+		 * @return a BufferResponse object with the header parsed and the data buffer to fill in later. The
+		 *			data buffer will be null if the target channel has been released or the buffer size is 0.
+		 */
+		static BufferResponse readFrom(ByteBuf messageHeader, NetworkBufferAllocator bufferAllocator) {
+			InputChannelID receiverId = InputChannelID.fromByteBuf(messageHeader);
+			int sequenceNumber = messageHeader.readInt();
+			int backlog = messageHeader.readInt();
+			boolean isBuffer = messageHeader.readBoolean();
+			boolean isCompressed = messageHeader.readBoolean();
+			int size = messageHeader.readInt();
 
-			ByteBuf retainedSlice = buffer.readSlice(size).retain();
-			return new BufferResponse(retainedSlice, isBuffer, isCompressed, sequenceNumber, receiverId, backlog);
+			Buffer dataBuffer = null;
+
+			if (size != 0) {
+				if (isBuffer) {
+					dataBuffer = bufferAllocator.allocatePooledNetworkBuffer(receiverId);
+				} else {
+					dataBuffer = bufferAllocator.allocateUnPooledNetworkBuffer(size);
+				}
+			}
+
+			if (dataBuffer != null) {
+				dataBuffer.setCompressed(isCompressed);
+			}
+
+			return new BufferResponse(
+				dataBuffer,
+				isBuffer,
+				isCompressed,
+				sequenceNumber,
+				receiverId,
+				backlog,
+				size);
 		}
 	}
 
 	static class ErrorResponse extends NettyMessage {
 
-		private static final byte ID = 1;
+		static final byte ID = 1;
 
 		final Throwable cause;