@pnowojski 2019-03-04
@gaoyunhaii Could you provide benchmark results (https://github.com/dataArtisans/flink-benchmarks/) to motivate this change?