@gaoyunhaii 2019-03-05
OK, I will complement the results. Besides, a more obvious advantage of receiver-side zero-copy is that it decreases the memory consumed by Netty, I will also complement the results on the memory consumption.