@NicoK 2019-06-06
Hi @gaoyunhaii and @pnowojski,
I took the liberty to rebase this PR on top of the latest master (see https://github.com/NicoK/flink/tree/f10742) and run our benchmarks. Unfortunately, changes are not really visible there (still, the reduced memory is also an argument for this PR). Slight improvements, maybe also in the error...

```
Benchmark                         Score±Error (master vs. PR)        
networkBroadcastThroughput          817 ±  24 vs.   832 ±  17
networkThroughput, 100,100ms      65096 ±2273 vs. 65451 ±1020
networkThroughput, 100,100ms,SSL  16432 ± 701 vs. 16748 ± 448
networkThroughput, 1000,1ms       41128 ± 903 vs. 42872 ± 934
networkThroughput, 1000,100ms     46166 ±1343 vs. 46304 ±1938
networkThroughput, 1000,100ms,SSL 13677 ± 173 vs. 13801 ± 196
```