@gaoyunhaii 2019-06-06
Hi @NicoK @pnowojski , very sorry for the long delay to response, and thanks a lot for helping me on the testing. I'm also still working on this PR:

1. For the throughput, I have the same test result with Nico's results. The PR has slight improvement on the throughput. Therefore, I'm keep trying to figure out if there may be some optimization on the implementation. 
2. However, I think the direct memory consumption reduction will be high. I'm also trying to add a new benchmark test on the direct memory to the flink-benchmark project to reveal it officially. 
3. At last, I'm also trying to split the big commit into multiple small ones for better review.

On our side, we have taken use of this PR to reduce the memory consumption and avoid OOM problem that occurs online frequently. I think I can  post the manual test on the memory consumption first in this week. 

IMO I think it will still be helpful by optimizing the memory footprint and achieving some performance improvement (although only slight improvement currently, and I'll try to improve that). How do you feel about that ?