@gaoyunhaii 2019-06-12
@NicoK Sorry since the memory test takes more time than expected. I have attached the main result in the following, and described the results in the [Google Doc](https://docs.google.com/document/d/1EGxlkCC-ZWNd5E7XVKWJEy3SbksI_XVDvZMciWFtOr8/edit?usp=sharing) with more details. As a whole, I think the test demonstrates that the modification could reduce 80% of direct memory used by Netty and reduce the risk of OOM.

![netty-memory](https://user-images.githubusercontent.com/1683890/59339911-a1df6280-8d37-11e9-9532-f6e76a3dab2c.png)