@gaoyunhaii 2019-06-14
Hi all, for the throughput, I will continue trying to improve the current implementation in the next few days and I'll try to have a conclusion as soon as possible.