@NicoK 2019-06-16
Thanks @gaoyunhaii for the in-depth analysis: these are very good results indeed - unfortunately, I won't have time next week to review this PR but will try to schedule for the week after if there is no other review until then.