@gaoyunhaii 2020-02-06
Very thanks @zhijiangW for the review! After the offline discussion with @zhijiangW , we come to the consistency that the reduction of memory should be mainly resulted from reducing the copy of data buffer in the received side, therefore it should be reasonable to first concentrate on this issue to reduce the complexity of the implementation. If we found that it is also necessary to reduce the consumption of memory produced by the header part of the message in the future, we could then continue optimizing this part, and by then we could also take the memory consumed by sending messages into consideration.