@gaoyunhaii 2020-03-10
The result of running performance benchmark for three times is as follows ant it shows the performance should be nearly unchanged after this PR.


Benchmark | (channelsFlushTimeout) | Master | PR
-- | -- | -- | --
DataSkewStreamNetworkThroughputBenchmarkExecutor.networkSkewedThroughput | N/A | 18544.274 ± 394.04 | 18693.114 ± 450.411
StreamNetworkBroadcastThroughputBenchmarkExecutor.networkBroadcastThroughput | N/A | 558.942 ± 13.561 | 572.935 ± 11.449
StreamNetworkThroughputBenchmarkExecutor.networkThroughput | 100,100ms | 39126.475 ± 5596.933 | 39598.01 ± 5355.437
StreamNetworkThroughputBenchmarkExecutor.networkThroughput | 100,100ms,SSL | 8011.816 ± 85.928 | 7988.84 ± 77.926
StreamNetworkThroughputBenchmarkExecutor.networkThroughput | 1000,1ms | 22692.533 ± 625.885 | 22510.761 ± 569.481
StreamNetworkThroughputBenchmarkExecutor.networkThroughput | 1000,100ms | 22412.451 ± 841.211 | 22157.080 ± 687.059
StreamNetworkThroughputBenchmarkExecutor.networkThroughput | 1000,100ms,SSL | 6192.704 ± 148.309 | 6305.707 ± 122.907
StreamNetworkLatencyBenchmarkExecutor.networkLatency1to1 | N/A | 10.056 ± 0.003 | 10.057 ± 0.005




Benchmark | (channelsFlushTimeout) | Master | PR
-- | -- | -- | --
DataSkewStreamNetworkThroughputBenchmarkExecutor.networkSkewedThroughput | N/A | 18227.444 ± 676.322 | 19311.283 ± 727.86
StreamNetworkBroadcastThroughputBenchmarkExecutor.networkBroadcastThroughput | N/A | 577.11 ± 11.601 | 572.644 ± 5.273
StreamNetworkThroughputBenchmarkExecutor.networkThroughput | 100,100ms | 38131.351 ± 6042.85 | 38395.832 ± 5891.265
StreamNetworkThroughputBenchmarkExecutor.networkThroughput | 100,100ms,SSL | 7945.597 ± 85.232 | 8016.364 ± 75.038
StreamNetworkThroughputBenchmarkExecutor.networkThroughput | 1000,1ms | 22725.854 ± 556.134 | 22044.464 ± 1357.626
StreamNetworkThroughputBenchmarkExecutor.networkThroughput | 1000,100ms | 21255.556 ± 818.414 | 21796.35 ± 808.031
StreamNetworkThroughputBenchmarkExecutor.networkThroughput | 1000,100ms,SSL | 6301.528 ± 146.149 | 6315.512 ± 125.690
StreamNetworkLatencyBenchmarkExecutor.networkLatency1to1 | N/A | 10.056 ± 0.003 | 10.056 ± 0.004



Benchmark | (channelsFlushTimeout) | Master | PR
-- | -- | -- | --
DataSkewStreamNetworkThroughputBenchmarkExecutor.networkSkewedThroughput | N/A | 17503.326 ± 1368.363 | 18014.828 ± 360.011
StreamNetworkBroadcastThroughputBenchmarkExecutor.networkBroadcastThroughput | N/A | 584.573 ± 7.798 | 546.446 ± 26.501
StreamNetworkThroughputBenchmarkExecutor.networkThroughput | 100,100ms | 37484.842 ± 4556.398 | 38647.774 ± 4133.883
StreamNetworkThroughputBenchmarkExecutor.networkThroughput | 100,100ms,SSL | 7985.803 ± 74.179 | 8035.338 ± 76.032
StreamNetworkThroughputBenchmarkExecutor.networkThroughput | 1000,1ms | 20226.816 ± 1852.461 | 22805.671 ± 719.033
StreamNetworkThroughputBenchmarkExecutor.networkThroughput | 1000,100ms | 21195.998 ± 1363.098 | 22430.819 ± 718.735
StreamNetworkThroughputBenchmarkExecutor.networkThroughput | 1000,100ms,SSL | 6220.015 ± 154.263 | 6176.15 ± 144.825
StreamNetworkLatencyBenchmarkExecutor.networkLatency1to1 | N/A | 10.056 ± 0.003 | 10.056 ± 0.003